Revert to "buildandpush" task in pipeline to push final container image.
This is not ideal as we're rebuilding after testing a built image.
However, we've had issues with the Docker tasks that make it difficult to build, use and push
the same image. Also, Docker should simply use cached layers for the entire rebuild since nothing
will have changed in the source code between the two build steps.